Updated pointing model for finders.
I've updated the pointing model for the finders. Last night, we saw big corrections for the finders which were almost immediately counteracted by C1. Perhaps we have the sense of the correction wrong? So, I'm now running the model the other way (using the finders as reference rather than C1).