Work around SI bug.
There is a bug in the firmware of the PCIe card that causes the blue PC to hang if an exposure is aborted. See email with <dgilmore@specinst.com> in 2020/2021. The best thing we can do in this circumstance is simply wait for the exposure to end and then read it out.