Implement "utility" images
Where appropriate, use "focus" and "astrometry" exposure types instead of "object" exposure types. These image types are written with a "u" suffix (for "utility") and do not create data cubes. The idea is to generate less useless data (data cubes) and to mark images that are not worth downloading (utility images). Implements #24.